refactor(bns): use OpenZeppelin upgrades deployment
