+++ src/apps/bns/package.json
@@ -10,16 +10,19 @@
     "test": "hardhat test solidity",
     "abi:sync": "node hardhat-scripts/sync-abi.mjs",
     "abi:check": "node hardhat-scripts/sync-abi.mjs --check",
-    "deploy:op-mainnet": "hardhat run hardhat-scripts/deploy-op-mainnet.ts --network opMainnet",
+    "validate:upgrade": "hardhat run hardhat-scripts/validate-upgrade.ts",
+    "deploy": "hardhat run hardhat-scripts/deploy.ts",
+    "deploy:op-mainnet": "hardhat run hardhat-scripts/deploy.ts --network opMainnet",
     "run:anvil": "hardhat run --network anvil"
   },
   "devDependencies": {
-    "@nomicfoundation/hardhat-viem": "3.0.9",
+    "@nomicfoundation/hardhat-ethers": "4.0.14",
     "@openzeppelin/contracts": "5.6.1",
     "@openzeppelin/contracts-upgradeable": "5.6.1",
+    "@openzeppelin/hardhat-upgrades": "4.0.2",
     "@types/node": "^22.13.0",
+    "ethers": "6.17.0",
     "hardhat": "^3.9.1",
-    "typescript": "^5.9.0",
-    "viem": "2.55.1"
+    "typescript": "^5.9.0"
   }
 }
